fixed imports and missing call
